requirements.txt
aiohttp==3.11.10
ddddocr==1.5.6
Flask==3.1.1
numpy==2.3.1
opencv_python_headless==4.10.0.84
Pillow==11.2.1
Requests==2.32.4
scikit_learn==1.7.0
